requirements.txt
cassandra-driver==3.12.0
certifi==2018.1.18
cffi==1.11.2
chardet==3.0.4
cql==1.4.0
falcon==1.4.1
greenlet==0.4.12
gunicorn==19.7.1
httpie==0.9.9
idna==2.6
msgpack==0.5.1
Pygments==2.2.0
python-mimeparse==1.6.0
readline==6.2.4.1
requests==2.18.4
serpy==0.3.1
six==1.11.0
thrift==0.11.0
urllib3==1.22
